+++ Lilu/Sources/kern_user.cpp
@@ -1152,15 +1152,18 @@ void UserPatcher::activate() {
 
 const char *UserPatcher::getSharedCachePath() {
 	bool isHaswell = BaseDeviceInfo::get().cpuHasAvx2;
-	if (getKernelVersion() >= KernelVersion::BigSur)
+	if (getKernelVersion() >= KernelVersion::Ventura)
+		return isHaswell ? venturaSharedCacheHaswell : venturaSharedCacheLegacy;
+	else if (getKernelVersion() >= KernelVersion::BigSur)
 		return isHaswell ? bigSurSharedCacheHaswell : bigSurSharedCacheLegacy;
 	return isHaswell ? sharedCacheHaswell : sharedCacheLegacy;
 }
 
 bool UserPatcher::matchSharedCachePath(const char *path) {
 	if (getKernelVersion() >= KernelVersion::BigSur) {
-		auto len = strlen(bigSurSharedCacheLegacy);
-		if (strncmp(path, bigSurSharedCacheLegacy, len) != 0)
+		auto dyld_path = getKernelVersion() >= KernelVersion::Ventura ? venturaSharedCacheLegacy : sharedCacheLegacy;
+		auto len = strlen(dyld_path);
+		if (strncmp(path, dyld_path, len) != 0)
 			return false;
 		path += len;
 	} else {